fix: actions and buttons
