+++ apps/api-journeys-modern/src/schema/blocks/action/action.zod.ts
@@ -1,0 +1,11 @@
+import { z } from 'zod'
+
+import { BlockSchema } from '../blocks.zod'
+
+const ActionSchema = z.object({
+  parentBlockId: z.string(),
+  parentBlock: BlockSchema,
+  gtmEventName: z.string().nullable()
+})
+
+export { ActionSchema }
